requirements.txt
# pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
--index-url https://pypi.org/simple/
--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
torchmetrics==1.0.3
lightning==2.4.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
rich==13.9.4
einops==0.8.0
rootutils==1.0.7
scikit-learn==1.6.1
scipy==1.15.2
huggingface_hub==0.26.5

notebook==6.1.4

# reading geo files
xarray==2025.1.2
dask==2025.1.0
rasterio==1.4.1
rioxarray==0.18.2
zarr==2.18.3
geopandas==1.0.1

# plotting
matplotlib==3.10.0
seaborn==0.13.2

# minicube creation/downloading
deap==1.4.2
terragon-downloader==0.1.1
sentinelhub==3.11.1
earthengine-api==1.5.8
boto3==1.35.49
botocore==1.35.49
rasterstats==0.20.0

# optional
joblib==1.4.2
optuna==4.1.0
tqdm==4.66.5
wandb==0.19.8
